feat: add tear_down_simulate() (#105)

+++ src/backend.c
@@ -94,7 +94,7 @@ parse_command (char *str, object_t * ob)
 
 /*  backend()
  */
-int eval_cost;
+size_t eval_cost;
 
 void
 backend ()
